requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.26
passlib==1.7.4
fastapi-users==9.1.1
fastapi-users[sqlalchemy]==9.1.1
psycopg2==2.9.9
PyJWT==2.3.0
boto3==1.34.82
cachetools==5.3.3
bcrypt==4.1.2
starlette==0.14.2
aiofiles==23.2.1
pytest==8.1.1